chore: split composer api
